a_event {r2vr}	R Documentation

a_event

Description

Build an A-Frame Event Message

Usage

a_event(id, event_name, event_detail = "", bubbles = FALSE)

Arguments

id
the id of the entity that will have the event emitted on it.

event_name
the name of the event.

event_detail
data associated with the event.

bubbles
TRUE if the event should bubble up to parents of the entity with 'id'?

Details

Construct an A-Frame event to be sent to a scene in browser, using a websocket connection. When the event is received by a scene with an active websocket server, it will be routed to the entity specified by 'id'. To handle the event the entity needs to have components that attach appropriate event handlers.

Value

An object that represents an A-Frame event.

[Package r2vr version 0.0.0.9000 Index]