a_kill_all_scenes {r2vr}	R Documentation

a_kill_all_scenes

Description

Kill all running A-Frame Scenes

Usage

a_kill_all_scenes()

Details

Frequently in development you may accidentally loose a handle to a running scene, meaning it will block a port you wish to run a new scene on. This kills all running scenes to free up all ports being used to serve by r2vr.

Value

nothing.

[Package r2vr version 0.0.0.9000 Index]