a_label {r2vr}	R Documentation

a_label

Description

Create a text label, visible in VR.

Usage

a_label(text, color = "#000000", font = "roboto", ...)

Arguments

text
the label text.

color
the label text colour.

font
the font to be used for the label text. Either the name of a built-in font, listed here: https://aframe.io/docs/0.8.0/components/text.html#stock-fonts, or a URL for an sdf .fnt file.

...
additional arguments to be passed to the A-Frame text entity, see: https://aframe.io/docs/0.8.0/primitives/a-text.html.

Details

This function returns an entity that will render a 2D text label in VR. It is wrapper for an A-Frame entity with the text component attached, see: https://aframe.io/docs/0.8.0/components/text.html. Additional component configuration arguments listed in the A-Frame documentation can be supplied in the '...' arguments.

Use standard entity components such as 'position', 'rotation', and 'scale' in '...' to place the label.

Value

an A-Frame entity object.

[Package r2vr version 0.0.0.9000 Index]