a_pc_control_camera {r2vr}	R Documentation

a_pc_control_camera

Description

Personal computer movement controls for an A-Frame scene camera

Usage

a_pc_control_camera(.js_sources = list(), easingY = 15,
  acceleration = 100, .version_num = "4.1.2", ...)

Arguments

.js_sources
Defaults to version 4.1.2 of the controls.

easingY
Easing rate of movement in the y direction (up).

acceleration
Acceleration of movement in forward, backward, left, and right directions.

.version_num
A-Frame extras version number.

...
additional components to be added to this entity.

Details

This function returns an entity that will enable the scene camera perspective to be controlled using WASD keys and the mouse.

The entity imports an external js source for controls written by Don McCurdy.

Value

An entity object describing a camera controlled by keyboard and mouse.

[Package r2vr version 0.0.0.9000 Index]